+++ .github/build-manylinux.sh
@@ -4,6 +4,7 @@ set -e -x
 cd /github/workspace/
 export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/src/dotnet/shared/Microsoft.NETCore.App/6.0.0-rc.1.21451.13
 echo $LD_LIBRARY_PATH
+rm -rf _skbuild/
 /opt/python/cp310-cp310/bin/pip install --upgrade --no-cache-dir pip
 /opt/python/cp310-cp310/bin/pip install auditwheel
 /opt/python/cp310-cp310/bin/pip install .