+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added support for generators in 3.10
 * Fixed a bug in the DUP_TOP_TWO opcode with escaped variables
 * UNPACK_SEQUENCE will emit types for constant tuples of string, int and float
+* Upgraded scikit-build to 0.12 for native VS2019 support
 
 ## 1.0.0 (rc3)