+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Added support for generators in 3.10
 * Fixed a bug in the DUP_TOP_TWO opcode with escaped variables
 * UNPACK_SEQUENCE will emit types for constant tuples of string, int and float
+* STORE_FAST is now profiled during the probe stage of PGC, leading to drastically improved performance in many cases where unboxed ints and floats can be used
+* Added a simple benchmark suite script (`Tests/benchmarks/suite.py`)
 * Upgraded scikit-build to 0.12 for native VS2019 support
 
 ## 1.0.0 (rc3)