+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release notes
 
 
-## 1.0.0 (develop/main)
+## 1.0.0 (rc4)
 
 * Added ARM64 for Apple macOS (M1 Silicon). The JIT will emit native ARM64 assembly. Disassembler not supported.
 * All JIT configuration is now set using the `pyjion.config()` function instead of `enable_feature()`/`disable_feature()`